refactor: rename dirs for clarity
