--- packages/slinkity/plugin/reactPlugin/1-pluginDefinitions/index.js
@@ -1,7 +1,0 @@
-const addShortcode = require('./addShortcode')
-const addPageExtension = require('./addPageExtension')
-
-module.exports = {
-  addShortcode,
-  addPageExtension,
-}
