+++ packages/slinkity/utils/toFormattedDataForProps.js
